Fix error
error: conversion to 'uint64_t' {aka 'long unsigned int'} from 'long long int' may change the sign of the result [-Werror=sign-conversion]

+++ link/.local/include/clmum.hpp
@@ -36,21 +36,21 @@ static inline uint64_t
 clmumx(__m128i a)
 {
 	clmul(a);
-	return _mm_extract_epi64(a, 1) ^ _mm_extract_epi64(a, 0);
+	return static_cast<uint64_t>(_mm_extract_epi64(a, 1)) ^ static_cast<uint64_t>(_mm_extract_epi64(a, 0));
 }
 
 static inline uint64_t
 clmuma(__m128i a)
 {
 	clmul(a);
-	return _mm_extract_epi64(a, 1) + _mm_extract_epi64(a, 0);
+	return static_cast<uint64_t>(_mm_extract_epi64(a, 1)) + static_cast<uint64_t>(_mm_extract_epi64(a, 0));
 }
 
 static inline uint64_t
 clmums(__m128i a)
 {
 	clmul(a);
-	return _mm_extract_epi64(a, 1) - _mm_extract_epi64(a, 0);
+	return static_cast<uint64_t>(_mm_extract_epi64(a, 1)) - static_cast<uint64_t>(_mm_extract_epi64(a, 0));
 }
 
 static void
@@ -60,8 +60,8 @@ clmul(uint64_t& hi, uint64_t& lo)
 
 	clmul(result);
 
-	hi = _mm_extract_epi64(result, 1);
-	lo = _mm_extract_epi64(result, 0);
+	hi = static_cast<uint64_t>(_mm_extract_epi64(result, 1));
+	lo = static_cast<uint64_t>(_mm_extract_epi64(result, 0));
 }
 
 static inline uint64_t
